refactor: reducer 함수분리
